Matmul: hypercube configs test (#748)
* cube count sm

* work by removing swizzle

* wip

* refactor into enums

* wip

* wip

* works

* comptime check

* cleanup and api

* cleanup and fix

* cleanup

* rename overallocate to ratio

* ci

* plane accelerated works

* more testing

* fix tma

* cleanup

* hypercube tests

* Add TODO for quantized matmul tests

* Apply review comments (PR #746)

---------

+++ crates/cubecl/benches/matmul.rs
@@ -271,6 +271,6 @@ fn main() {
     }
     #[cfg(feature = "wgpu-msl")]
     {
-        run_benches::<cubecl::wgpu::WgpuRuntime, f32>();
+        run_benches::<cubecl::wgpu::WgpuRuntime, half::f16>();
     }
 }
